requirements.txt
Flask==3.0.3
Flask-Mail==0.10.0
authlib==1.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
bcrypt==4.1.3